changing the files up (index with server)
