+++ server/src/index.ts
@@ -1,18 +1,23 @@
 import express, { Application } from "express";
-import cors, { CorsOptions } from "cors";
+import Server from "./server";
+import userRouter from "./routes/user.routes";
+import catalogRouter from "./routes/catalog.routes";
 
-export default class Server {
-  constructor(app: Application) {
-    this.config(app);
-  }
+const app: Application = express();
+const server: Server = new Server(app);
+const PORT: number = process.env.PORT ? parseInt(process.env.PORT, 10) : 8080;
 
-  private config(app: Application): void {
-    const corsOptions: CorsOptions = {
-      origin: "http://localhost:5173",
-    };
+app.use("/api/users", userRouter);
+app.use("/api/catalog", catalogRouter);
 
-    app.use(cors(corsOptions));
-    app.use(express.json());
-    app.use(express.urlencoded({ extended: true }));
-  }
-}
+app
+  .listen(PORT, "localhost", function () {
+    console.log(`Server is running on port ${PORT}.`);
+  })
+  .on("error", (err: any) => {
+    if (err.code === "EADDRINUSE") {
+      console.log("Error: address already in use");
+    } else {
+      console.log(err);
+    }
+  });
