+++ server/src/server.ts
@@ -1,23 +1,18 @@
 import express, { Application } from "express";
-import Server from "./index";
-import userRouter from "./routes/user.routes";
-import catalogRouter from "./routes/catalog.routes";
+import cors, { CorsOptions } from "cors";
 
-const app: Application = express();
-const server: Server = new Server(app);
-const PORT: number = process.env.PORT ? parseInt(process.env.PORT, 10) : 8080;
+export default class Server {
+  constructor(app: Application) {
+    this.config(app);
+  }
 
-app.use("/api/users", userRouter);
-app.use("/api/catalog", catalogRouter);
+  private config(app: Application): void {
+    const corsOptions: CorsOptions = {
+      origin: "http://localhost:5173",
+    };
 
-app
-  .listen(PORT, "localhost", function () {
-    console.log(`Server is running on port ${PORT}.`);
-  })
-  .on("error", (err: any) => {
-    if (err.code === "EADDRINUSE") {
-      console.log("Error: address already in use");
-    } else {
-      console.log(err);
-    }
-  });
+    app.use(cors(corsOptions));
+    app.use(express.json());
+    app.use(express.urlencoded({ extended: true }));
+  }
+}
