+++ server/vercel.json
@@ -1,0 +1,3 @@
+{
+  "rewrites": [{ "source": "/api/(.*)", "destination": "/api" }]
+}
